requirement.txt
absl-py==1.2.0
cachetools==5.2.0
charset-normalizer==2.1.1
compressai==1.2.1
contourpy==1.0.5
cycler==0.11.0
einops==0.5.0
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
importlib-metadata==5.0.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib
numpy
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==0.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.1
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
